requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.10
idna==3.3
requests==2.27.0
urllib3==1.26.7
uuid==1.30
paho-mqtt==1.6.1
python-dateutil==2.8.2
pyyaml==6.0
